some refactoring

+++ etl/csv_to_db.py
@@ -10,8 +10,9 @@ def etl_main(db_manager,
              do_extract=False,
              ds_owner='svanoo',
              ds_name='myanimelist-dataset',
-             data_dir='O:/dev/_ml/mal_stats/data/svanoo',
+             data_dir='./data/',
              user_fields=['user_id', 'user_url'],
+             item_fields=['anime_id', 'anime_url', 'title', 'main_pic'],
              rating_fields=['user_id', 'anime_id', 'score', 'status']):
 
     # Extract all
@@ -27,30 +28,49 @@ def etl_main(db_manager,
         logging.info('Dataset already extracted')
 
     # Transform and load users
-    user_df = pd.read_csv(os.path.join(
-        data_dir, 'user.csv'), delimiter='\t', usecols=user_fields)
-    user_df.rename(columns={'user_id': 'user_name'}, inplace=True)
-    user_df = user_df.rename_axis('user_id').reset_index()
-    verror = db_manager.pandas_to_sql(
-        user_df, 'mal_user', if_exists='replace')
-    logging.info(verror if verror else "'mal_user' table processed and loaded")
+    if user_fields:
+        user_df = pd.read_csv(os.path.join(
+            data_dir, 'user.csv'), delimiter='\t', usecols=user_fields)
+        user_df.rename(columns={'user_id': 'user_name'}, inplace=True)
+        user_df = user_df.rename_axis('user_id').reset_index()
+        db_manager.drop_table('mal_user')
+        verror = db_manager.pandas_to_sql_bulk_postgres(user_df, 'mal_user')
+        logging.info(
+            verror if verror else "'mal_user' table processed and loaded")
+    else:
+        logging.info('No user fields to process')
+
+    # Transform and load items
+    if item_fields:
+        item_df = pd.read_csv(os.path.join(
+            data_dir, 'anime.csv'), delimiter='\t', usecols=item_fields)
+        db_manager.drop_table('mal_item')
+        verror = db_manager.pandas_to_sql_bulk_postgres(item_df, 'mal_item')
+        logging.info(
+            verror if verror else "'mal_item' table processed and loaded")
+    else:
+        logging.info('No item fields to process')
 
     # Transform and load ratings
-    rating_paths = [path for path in os.listdir(
-        data_dir) if 'user_anime' in path]
-    db_manager.drop_table('mal_rating')
-    for chunk_id, path in enumerate(rating_paths):
-        rating_df = pd.read_csv(os.path.join(data_dir, path),
-                                delimiter='\t',
-                                usecols=rating_fields)
-        rating_df = rating_df.loc[rating_df['status'] == 'completed'].dropna()
-        rating_df.rename(columns={'user_id': 'user_name'}, inplace=True)
-        rating_df = rating_df.astype({'score': int})
-        rating_df = rating_df.join(
-            user_df.set_index('user_name'), on='user_name')[['user_id', 'anime_id', 'score']]
-        verror = db_manager.pandas_to_sql_bulk_postgres(
-            rating_df, 'mal_rating')
-        logging.info(
-            verror if verror else "'mal_rating' chunk #{} processed and loaded".format(chunk_id))
-        del rating_df
-        gc.collect()
+    if rating_fields:
+        rating_paths = [path for path in os.listdir(
+            data_dir) if 'user_anime' in path]
+        db_manager.drop_table('mal_rating')
+        for chunk_id, path in enumerate(rating_paths):
+            rating_df = pd.read_csv(os.path.join(data_dir, path),
+                                    delimiter='\t',
+                                    usecols=rating_fields)
+            rating_df = rating_df.loc[rating_df['status']
+                                      == 'completed'].dropna()
+            rating_df.rename(columns={'user_id': 'user_name'}, inplace=True)
+            rating_df = rating_df.astype({'score': int})
+            rating_df = rating_df.join(
+                user_df.set_index('user_name'), on='user_name')[['user_id', 'anime_id', 'score']]
+            verror = db_manager.pandas_to_sql_bulk_postgres(
+                rating_df, 'mal_rating')
+            logging.info(
+                verror if verror else "'mal_rating' chunk #{} processed and loaded".format(chunk_id))
+            del rating_df
+            gc.collect()
+    else:
+        logging.info('No rating fields to process')
